remove dollar sign; it's not part of the command
GitHub uses a `clipboard-copy` button, so instead of highlighting text to copy and paste everything except the dollar sign, users can simply click the "Copy to Clipboard" button. Unfortunately, clicking this button also copies the dollar sign. Pasting this copied text into a terminal results in an error like `command not found: $`

Let's remove the dollar sign from the copyable text, and instead use another way to denote that this "code block" is meant to be run inside of a terminal.